revise machine-readable notifications for activity stream
Revise machine-readable notifications for Activity Stream conformance. I have some qualms about the way I've used `content` in the error notification specification. It's not really supposed to be used in the way I have, it's supposed to provide a textual representation of the content of the object. However, I couldn't find a better place to record that information.